+++ test/record/classProperties/instantiateTest.js
@@ -62,7 +62,7 @@ describe('Viking.Record::instantiate', () => {
         assert.equal(fleet.readAttribute('id'), 1);
         assert.equal(fleet.readAttribute('name'), 'Colonial Fleet');
         
-        let ships = await fleet.ships.toArray()
+        let ships = await fleet.ships.load()
         assert.equal(1, ships.length);
         assert.equal(ships[0].readAttribute('id'), 1);
         assert.equal(ships[0].readAttribute('name'), 'Galactica');